create new flattened_document_reference(); abstract into entities
